Add sounds, refactor

+++ Sensei/Classes/Managers/TutorialManager.swift
@@ -132,22 +132,22 @@ class TutorialManager {
         }
     }
     
-    func delayForCurrentStep() -> Double {
+    func delayForCurrentStep() -> Float {
         if !TutorialManager.sharedInstance.completed {
             if let currentStep = TutorialManager.sharedInstance.currentStep {
                 if let prevStep = TutorialManager.sharedInstance.prevTutorialStep {
-                    let delayBefore = (prevStep.text.characters.count == 0 || currentStep.delayBefore == 0) ? currentStep.delayBefore : Double(prevStep.text.characters.count) * 0.03
-                    return  delayBefore
+                    let delayBefore = (prevStep.text.characters.count == 0 || currentStep.delayBefore == 0) ? currentStep.delayBefore : Float(prevStep.text.characters.count) * 0.03
+                    return  Float(delayBefore)
                 }
-                return currentStep.delayBefore
+                return Float(currentStep.delayBefore)
             }
         } else {
             if let currentStep = TutorialManager.sharedInstance.currentUpgradedStep {
                 if let prevStep = TutorialManager.sharedInstance.prevUpgradedStep {
-                    let delayBefore = (prevStep.text.characters.count == 0 || currentStep.delayBefore == 0) ? currentStep.delayBefore : Double(prevStep.text.characters.count) * 0.03
-                    return delayBefore
+                    let delayBefore = (prevStep.text.characters.count == 0 || currentStep.delayBefore == 0) ? currentStep.delayBefore : Float(prevStep.text.characters.count) * 0.03
+                    return  Float(delayBefore)
                 }
-                return currentStep.delayBefore
+                return Float(currentStep.delayBefore)
             }
         }
         return 0
